create utils file
this is mostly to reduce the unrelated diff on https://github.com/mmkal/expect-type/pull/83
